+++ app/serializers/pathfinder2/character_serializer.rb
@@ -6,14 +6,16 @@ module Pathfinder2
 
     attributes :features, :id, :name, :level, :race, :subrace, :main_class, :classes, :languages, :health, :abilities, :money,
                :skills, :created_at, :subclasses, :background, :saving_throws_value, :saving_throws, :dying_condition_value,
-               :avatar, :boosts, :weapon_skills, :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions,
-               :ability_boosts_v2, :skill_boosts, :attacks, :provider, :lores, :selected_features
+               :avatar, :weapon_skills, :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions,
+               :ability_boosts_v2, :skill_boosts, :attacks, :provider, :lores, :selected_features, :spells_info, :class_dc,
+               :spell_attack, :spell_dc, :spent_spell_slots
 
-    delegate :id, :name, :level, :race, :subrace, :main_class, :classes, :languages, :health, :abilities, :skills, :subclasses,
-             :background, :saving_throws_value, :saving_throws, :dying_condition_value, :boosts, :weapon_skills, :money, :lores,
-             :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions, :attacks, :features, to: :decorator
-    delegate :data, :created_at, to: :object
-    delegate :ability_boosts_v2, :skill_boosts, :selected_features, to: :data
+    delegate :health, :abilities, :skills, :subclasses, :saving_throws_value, :dying_condition_value, :weapon_skills,
+             :money, :lores, :armor_skills, :coins, :load, :armor_class, :speed, :perception, :conditions, :attacks,
+             :features, :spells_info, :class_dc, :spell_attack, :spell_dc, to: :decorator
+    delegate :name, :id, :data, :created_at, to: :object
+    delegate :level, :ability_boosts_v2, :skill_boosts, :selected_features, :race, :subrace, :main_class, :classes, :languages,
+             :background, :saving_throws, :spent_spell_slots, to: :data
 
     def provider
       'pathfinder2'